Move WebViewUtils into BrowserUtils
