+++ zygote/src/main/java/org/frknkrc44/hma_oss/zygote/util/BrowserUtils.kt
@@ -1,10 +1,17 @@
 package org.frknkrc44.hma_oss.zygote.util
 
 import android.os.Build
+import android.os.ServiceManager
+import android.provider.Settings
+import android.webkit.IWebViewUpdateService
 import com.android.server.pm.PackageManagerService
+import icu.nullptr.hidemyapplist.common.Utils.binderLocalScope
 import org.frknkrc44.hma_oss.zygote.service.HMAService.Companion.service
 import org.frknkrc44.hma_oss.zygote.util.Logcat.logD
+import org.frknkrc44.hma_oss.zygote.util.ServiceUtils.contentResolver
 import org.frknkrc44.hma_oss.zygote.util.ZLUtils.getObjectField
+import org.frknkrc44.hma_oss.zygote.util.ZygoteConstants.WEBVIEW_PROVIDER_KEY
+import org.frknkrc44.hma_oss.zygote.util.ZygoteConstants.WEBVIEW_UPDATE_SERVICE
 
 object BrowserUtils {
     const val TAG = "BrowserUtils"
@@ -15,6 +22,18 @@ object BrowserUtils {
         } catch (e: Throwable) {
             logD(TAG, e) { "Getting default browser failed" }
             null
+        }
+    }
+
+    fun getWebviewProvider(): String? = binderLocalScope {
+        try {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+                webViewService.currentWebViewPackage?.packageName
+            } else {
+                webViewService.currentWebViewPackageName
+            }
+        } catch (_: Throwable) {
+            Settings.Global.getString(contentResolver, WEBVIEW_PROVIDER_KEY)
         }
     }
 
@@ -29,5 +48,9 @@ object BrowserUtils {
             Build.VERSION_CODES.R -> pms.getPermissionManagerServiceInternal().getDefaultBrowser(userId)
             else -> pms.defaultAppProvider.getDefaultBrowser(userId)
         }
+    }
+
+    private val webViewService by lazy {
+        ServiceManager.getService(WEBVIEW_UPDATE_SERVICE) as IWebViewUpdateService
     }
 }
